Moved types to related components folders
