+++ static/client/components/Products/Products.types.ts
@@ -1,0 +1,8 @@
+import type { MultiSelectItem } from "@canonical/react-components";
+
+import type { IPage } from "@/services/api/types/pages";
+
+export interface IProductsProps {
+  page?: IPage;
+  onSelectProducts?: (products: MultiSelectItem[]) => void;
+}
