+++ static/client/components/Reporter/Reporter.types.ts
@@ -1,0 +1,6 @@
+import type { IUser } from "@/services/api/types/users";
+
+export interface IReporterProps {
+  reporter: IUser | null;
+  setReporter: React.Dispatch<React.SetStateAction<IUser>>;
+}
